backend/requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib 
python-dotenv
 
opik
langchain
langchain-community
langchain-core
langchain-openai
langgraph
langchain-google-genai

langchain-groq
langchain-mistralai

fastapi "uvicorn[standard]" 
python-multipart 
sqlalchemy 
aiosqlite 
requests

annotated-types 
pydantic 
pydantic-core

uvicorn
